requirements.txt
streamlit>=1.32.0
scikit-learn>=1.2.0
Pillow>=9.0.0
numpy>=1.23.0
joblib>=1.2.0
opencv-python>=4.7.0
